Crypto: Workaround: fixed crashing user linked list
This fixes bug #169
This is only a workaround. There has to be a new storage possibility for
users, searchable by their fingerprints, as well as a separated
network/IP/MAC view with all the nodes and how to reach them. The users
could then optionally be attached to such a node.